+++ src/sodiumoxide/crypto/edwards25519sha512batch.rs
@@ -106,9 +106,9 @@ pub fn verify(sm: &[u8],
 #[test]
 fn test_sign_verify() {
     use randombytes::randombytes;
-    for i in range(0, 256) {
+    for i in range(0, 256u) {
         let (pk, sk) = gen_keypair();
-        let m = randombytes(i as uint);
+        let m = randombytes(i);
         let sm = sign(m, &sk);
         let m2 = verify(sm, &pk);
         assert!(Some(m) == m2);
@@ -118,9 +118,9 @@ fn test_sign_verify() {
 #[test]
 fn test_sign_verify_tamper() {
     use randombytes::randombytes;
-    for i in range(0, 32) {
+    for i in range(0, 32u) {
         let (pk, sk) = gen_keypair();
-        let m = randombytes(i as uint);
+        let m = randombytes(i);
         let mut sm = sign(m, &sk);
         for j in range(0, sm.len()) {
             sm[j] ^= 0x20;
